// src/openapi/model/user.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    id: string;
    name: string;
    surname: string;
    email: string;
    password?: string;
    cellphone: string;
    user_type: string;
}